Moved env_options.py to apio/utils.
